# apis/api.yaml
openapi: 3.0.0
info:
  title: Canvara Backend API
  version: 0.11.0
servers: []
...
paths:
...
  /posts:
    get:
      summary: List all posts
      tags:
        - posts
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Post'
      operationId: list_posts
      security:
        - AWSCognitoCanvaraPrototype: []
      parameters:
        - schema:
            type: string
          in: query
          name: post_owner_id
          description: Unique identifier of the post owner.
        - schema:
            type: string
          in: query
          name: q
          description: The search query.
          deprecated: true
        - schema:
            type: string
            default: curated
          in: query
          name: filter
          description: Filter for the results.
          deprecated: true
        - schema:
            type: string
          in: query
          name: type
          description: 'Unique identifier of the post type for search for. If absent, return posts of all types.'
          deprecated: true
        - schema:
            type: string
            enum:
              - latest
              - recommended
            default: recommended
          in: query
          name: sort
          description: 'Sorting identifier for list of gigs.'
        - schema:
            type: string
          in: query
          name: keyword
          description: 'Search keyword for gigs.'
        - schema:
            type: string
            enum:
              - open
              - underway
              - completed
              - suspended
          in: query
          name: status
          description: 'Status filter for list of gigs.'
        - schema:
            type: string
            enum:
              - s
              - m
              - l
          in: query
          name: project_size
          description: 'Project size filter for list of gigs.'
        - schema:
            type: string
          in: query
          name: target_date
          description: 'Date filter upto target date.'
        - schema:
            type: string
          in: query
          name: location_id
          description: 'Location id filter for list of gigs.'
        - schema:
            type: string
          in: query
          name: department
          description: 'Department filter for list of gigs.'
    post:
      summary: Add a new post
      tags:
        - posts
      operationId: create_post
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Post'
      security:
        - AWSCognitoCanvaraPrototype: []
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                  description: Name of this post.
                task_owner_id:
                  type: string
                  deprecated: true
                post_type_id:
                  type: string
                  description: Unique identifier of the post type for this post.
                description:
                  type: string
                  description: Description of this post.
                size:
                  type: string
                  description: T-shirt size of this post.
                  enum:
                    - S
                    - M
                    - L
                target_date:
                  type: string
                  description: 'Target date for the completion of work described in this post, in ISO 8601 format.'
                expiration_date:
                  type: string
                  description: ISO 8601 formatted Date on which this post expires if no candidate is selected.
                language:
                  type: string
                  description: Language for this post.
                location_id:
                  type: string
                  description: Unique identifier of the location for this post.
                people_needed:
                  type: integer
                  description: Number of people needed for this post.
                required_skills:
                  type: array
                  description: Must-have skills for this post.
                  items:
                    $ref: '#/components/schemas/SkillWithLevel'
                desired_skills:
                  type: array
                  description: Nice-to-have skills for this post.
                  items:
                    $ref: '#/components/schemas/SkillWithLevel'
                candidate_description:
                  type: string
                  description: Description of an ideal candidate for this post.
                auto_assign:
                  type: boolean
                  description: Automatically select applications or not.
                canvara_kudos:
                  type: boolean
                  description: Canvara kudos true or false.
                hashtags:
                  type: array
                  items:
                    type: string
                  description: List of hastags or keywords.
                highlighted_communities:
                  type: array
                  description: Highlighted Community ids for gig.
                  items:
                    type: string
              required:
                - name
                - post_type_id
                - size
                - target_date
                - language
                - location_id
                - people_needed
  '/posts/{post_id}':
    parameters:
      - schema:
          type: string
        name: post_id
        in: path
        required: true
    get:
      summary: Get details of a post
      tags:
        - posts
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Post'
      operationId: get_post
      security:
        - AWSCognitoCanvaraPrototype: []
    put:
      summary: Update details of a post
      tags:
        - posts
      operationId: update_post
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Post'
      security:
        - AWSCognitoCanvaraPrototype: []
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                task_owner_id:
                  type: string
                  deprecated: true
                name:
                  type: string
                  description: Name of this post.
                post_type_id:
                  type: string
                  description: Unique identifier of the post type for this post.
                status:
                  type: string
                  description: Current status of this post.
                description:
                  type: string
                  description: Description of this post.
                size:
                  type: string
                  description: T-shirt size of this post.
                  enum:
                    - S
                    - M
                    - L
                target_date:
                  type: string
                  description: 'Target date for the completion of work described in this post, in ISO 8601 format.'
                expiration_date:
                  type: string
                  description: ISO 8601 formatted Date on which this post expires if no candidate is selected.
                language:
                  type: string
                  description: Language for this post.
                location_id:
                  type: string
                  description: Unique identifier of the location for this post.
                people_needed:
                  type: integer
                  description: Number of people needed for this post.
                required_skills:
                  type: array
                  description: Must-have skills for this post.
                  items:
                    $ref: '#/components/schemas/SkillWithLevel'
                desired_skills:
                  type: array
                  description: Nice-to-have skills for this post.
                  items:
                    $ref: '#/components/schemas/SkillWithLevel'
                candidate_description:
                  type: string
                  description: Description of an ideal candidate for this post.
                auto_assign:
                  type: boolean
                  description: Automatically select applications or not.
                canvara_kudos:
                  type: boolean
                  description: Canvara kudos true or false.
                hashtags:
                  type: array
                  items:
                    type: string
                  description: List of hastags or keywords.
                highlighted_communities:
                  type: array
                  description: Highlighted Community ids for gig.
                  items:
                    type: string
    delete:
      summary: Delete a post
      tags:
        - posts
      operationId: delete_post
      responses:
        '204':
          description: No Content
      security:
        - AWSCognitoCanvaraPrototype: []
...
  '/posts/{post_id}/video':
    parameters:
      - schema:
          type: string
        name: post_id
        in: path
        required: true
        description: Unique identifier of the post.
    put:
      summary: Create request for a post video upload
      operationId: upload_post_video
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserUploadResponse'
      tags:
        - posts
      security:
        - AWSCognitoCanvaraPrototype: []
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserUploadRequest'
            examples:
              Example:
                value:
                  filename: my_video.mp4
                  content_type: video/mp4
  '/posts/{post_id}/video/{upload_id}':
    parameters:
      - schema:
          type: string
        name: post_id
        in: path
        required: true
        description: Unique identifier of the post.
      - schema:
          type: string
        name: upload_id
        in: path
        required: true
        description: Unique identifier of the file upload.
    put:
      summary: Finalize post video upload
      operationId: finalize_post_video_upload
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
      tags:
        - posts
      security:
        - AWSCognitoCanvaraPrototype: []
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                status:
                  type: string
                  description: Status of the file upload.
                  enum:
                    - uploaded
              required:
                - status
    delete:
      summary: Delete a post video
      operationId: delete_post_video
      responses:
        '204':
          description: No Content
      tags:
        - posts
      security:
        - AWSCognitoCanvaraPrototype: []
  '/posts/{post_id}/like':
    parameters:
      - schema:
          type: string
        name: post_id
        in: path
        required: true
        description: Unique identifier of the post.
    put:
      summary: Add like for a post
      operationId: add_post_like
      responses:
        '204':
          description: No Content
      tags:
        - posts
      security:
        - AWSCognitoCanvaraPrototype: []
    delete:
      summary: Remove like for a post
      operationId: delete_post_like
      responses:
        '204':
          description: No Content
      tags:
        - posts
      security:
        - AWSCognitoCanvaraPrototype: []
  '/posts/{post_id}/bookmark':
    parameters:
      - schema:
          type: string
        name: post_id
        in: path
        required: true
        description: Unique identifier of the post.
    put:
      summary: Bookmark a post
      operationId: add_post_bookmark
      responses:
        '204':
          description: ''
      tags:
        - posts
      security:
        - AWSCognitoCanvaraPrototype: []
    delete:
      summary: Remove bookmark for a post
      operationId: delete_post_bookmark
      responses:
        '204':
          description: No Content
      tags:
        - posts
      security:
        - AWSCognitoCanvaraPrototype: []
...
  /onboarding/desired_skills:
    post:
      summary: Set desired skills during onboarding
      tags:
        - onboarding
      operationId: set_desired_skills
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Skill'
      security:
        - AWSCognitoCanvaraPrototype: []
      requestBody:
        content:
          application/json:
            schema:
              type: array
              items:
                $ref: '#/components/schemas/Skill'
...
  /post_types:
    get:
      summary: List all post types available to user
      tags:
        - posts
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/PostType'
      operationId: list_post_types
      security:
        - AWSCognitoCanvaraPrototype: []
  /locations:
    get:
      summary: List all post locations available to user
      tags:
        - posts
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Location'
      operationId: list_locations
      security:
        - AWSCognitoCanvaraPrototype: []
  /languages:
    get:
      summary: List all post languages available to user
      tags:
        - posts
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  type: string
      operationId: list_languages
      security:
        - AWSCognitoCanvaraPrototype: []
...
components:
  schemas:
...
    Post:
      title: Post
      type: object
      description: ''
      properties:
        customer_id:
          type: string
          deprecated: true
          description: Unique identifier of the customer to which this post belongs.
        post_id:
          type: string
          description: Unique identifier for this post.
        post_owner_id:
          type: string
          deprecated: true
          description: Unique identifier of the owner of this post.
        post_owner:
          type: object
          description: Information about the owner of this post.
          required:
            - user_id
            - name
          properties:
            user_id:
              type: string
              description: Unique identifier for the user.
            name:
              type: string
              description: Name of the user.
            profile_picture_url:
              type: string
              description: The URL of user's profile picture.
            location:
              type: string
              description: Location of the post owner.
        task_owner_id:
          type: string
          deprecated: true
        post_type:
          $ref: '#/components/schemas/PostType'
        name:
          type: string
          description: Name of this post.
        status:
          type: string
          description: Current status of this post.
        summary:
          type: string
          deprecated: true
          description: Name of this post.
        description:
          type: string
          description: Description of this post.
        video_url:
          type: string
          description: The URL of video description for this post.
        description_video:
          $ref: '#/components/schemas/UserUpload'
        size:
          type: string
          description: T-shirt size of this post.
          enum:
            - S
            - M
            - L
        location:
          $ref: '#/components/schemas/Location'
        people_needed:
          type: integer
          description: Number of people needed for this post.
        language:
          type: string
          description: Language for this post.
        required_skills:
          type: array
          description: Must-have skills for this post.
          items:
            $ref: '#/components/schemas/SkillWithLevel'
        desired_skills:
          type: array
          description: Nice-to-have skills for this post.
          items:
            $ref: '#/components/schemas/SkillWithLevel'
        candidate_description:
          type: string
          description: Description of an ideal candidate for this post.
        auto_assign:
          type: boolean
          description: Automatically select applications or not.
        canvara_kudos:
          type: boolean
          description: Canvara kudos true or false.
        hashtags:
          type: array
          items:
            type: string
          description: List of hastags or keywords.
        highlighted_communities:
          type: array
          description: Highlighted Community ids for gig.
          items:
            type: object
            description: Information about the community.
            required:
              - community_id
              - name
            properties:
              community_id:
                type: string
                description: Unique identifier for the community.
              name:
                type: string
                description: Name of the community.
              community_logo:
                type: string
                description: The URL of community logo.
        match_level:
          type: number
          description: A number between 0 and 100 indicating how close of a match this post is for the current user. Returned only for active posts and if the current user is not the post owner.
        matched_skills:
          type: array
          description: Post's skills that match the user's skills.
          items:
            $ref: '#/components/schemas/SkillWithLevel'
        unmatched_skills:
          type: array
          description: Post's skills that don't match the user's skills.
          items:
            $ref: '#/components/schemas/SkillWithLevel'
        is_bookmarked:
          type: boolean
          description: Whether this post has been bookmarked by the current user or not.
        is_liked:
          type: boolean
          description: Whether this post has been liked by the current user or not.
        target_date:
          type: string
          description: 'Target date for the completion of work described in this post, in ISO 8601 format.'
        expiration_date:
          type: string
          description: ISO 8601 formatted Date on which this post expires if no candidate is selected.
        created_at:
          type: string
          description: 'Date and time when this post was created, in ISO 8601 format.'
        last_updated_at:
          type: string
          description: 'Date and time when this post was last updated, in ISO 8601 format.'
      required:
        - post_id
        - post_owner
        - post_type
        - name
        - status
        - size
        - location
        - people_needed
        - language
        - target_date
...
    SkillWithLevel:
      title: SkillWithLevel
      allOf:
        - $ref: '#/components/schemas/Skill'
        - type: object
          properties:
            level:
              type: integer
...
    UserUploadResponse:
      title: UserUploadResponse
      type: object
      properties:
        upload_id:
          type: string
          description: Unique identifier of this upload.
        url:
          type: string
          description: Url where the file should be uploaded to.
      required:
        - upload_id
        - url
...
  securitySchemes:
    AWSCognitoCanvaraPrototype:
      type: http
      scheme: bearer